Fixed removing from pocket causing a UNIQUE constraint failure in some cases.
The issue is caused by the fact that removing an element from the pocket
causes values to be decremented one by one, therefor there are (temporarily)
multiple identical vlaues in the rank col.

I would like to find a way to keep the UNIQUE constraint, but I'd need to update the rows all at the same time,
and for now creating a temporary table to flush those values in seems like overkill.

THE CURRENT FIX is removing the UNIQUE constraint.